fix: harden geoip parser, probe miss-detection and CSV export (review)
Adversarial review of the six-feature batch found no critical/high defects;
these are the three low-severity robustness/correctness fixes it surfaced.

- geo: reject an over-long protobuf length varint by unsigned comparison — a
  length ≥ 2^63 cast to int went negative and slipped past the bounds check,
  panicking the slice on a corrupt/truncated geoip.dat (now errors cleanly)
- decoy: Miss() now mirrors serveMiss exactly — an extensionless miss on a
  single-page template is served as the index under 200, so it is NOT counted
  as a probe (only a 404-serving template or a missing asset is), preventing a
  future SPA-style decoy from flagging real visitors as scanners
- audit CSV: neutralize spreadsheet formula injection (leading = + - @ tab CR)
  in the free-text columns
- tests: malformed/truncated geoip.dat no-panic; Miss for SPA vs classic decoy

+++ internal/geo/lookup.go
@@ -47,7 +47,9 @@ func LoadCountryLookup(dir string) (*CountryLookup, error) {
 		}
 		data = data[n:]
 		msgLen, n := binary.Uvarint(data)
-		if n <= 0 || int(msgLen) > len(data[n:]) {
+		// Unsigned compare: a length varint ≥ 2^63 makes int(msgLen) negative, which
+		// would slip past a signed `> len` check and then panic on the slice.
+		if n <= 0 || msgLen > uint64(len(data[n:])) {
 			break
 		}
 		data = data[n:]
@@ -208,7 +210,9 @@ func readTag(msg []byte) (field uint64, wire byte, rest []byte, ok bool) {
 
 func readBytes(msg []byte) (b, rest []byte, ok bool) {
 	l, n := binary.Uvarint(msg)
-	if n <= 0 || int(l) > len(msg[n:]) {
+	// Unsigned compare so a huge (or overflow-to-negative-when-cast) length can't slip
+	// past the bound and panic the slice below.
+	if n <= 0 || l > uint64(len(msg[n:])) {
 		return nil, nil, false
 	}
 	return msg[n : n+int(l)], msg[n+int(l):], true
